refactor: WebController -> AppController 이름 변경
