+++ src/main/java/io/github/daengdaenglee/mangurl/inboundadapter/app/form/ShortenUrlForm.java
@@ -1,0 +1,4 @@
+package io.github.daengdaenglee.mangurl.inboundadapter.app.form;
+
+public record ShortenUrlForm(String originalUrl) {
+}
