--- src/main/java/io/github/daengdaenglee/mangurl/inboundadapter/web/form/ShortenUrlForm.java
@@ -1,4 +1,0 @@
-package io.github.daengdaenglee.mangurl.inboundadapter.web.form;
-
-public record ShortenUrlForm(String originalUrl) {
-}
